dom/html/test/forms/test_select_input_change_event.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1265968
-->
<head>
  <title>Test for Bug 1024350</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1024350">Mozilla Bug 1024350</a>
<p id="display"></p>
<div id="content">
  <select oninput='++selectInput;' onchange="++selectChange;">
    <option>one</option>
  </select>
  <select oninput='++selectInput;' onchange="++selectChange;">
    <option>one</option>
    <option>two</option>
  </select>
  <select multiple size='1' oninput='++selectInput;' onchange="++selectChange;">
    <option>one</option>
  </select>
  <select multiple oninput='++selectInput;' onchange="++selectChange;">
    <option>one</option>
    <option>two</option>
  </select>
</div>
<pre id="test">
<script type="application/javascript">
  var selectSingleOneItem = document.getElementsByTagName('select')[0];
  var selectSingle = document.getElementsByTagName('select')[1];
  var selectMultipleOneItem = document.getElementsByTagName('select')[2];
  var selectMultiple = document.getElementsByTagName('select')[3];

  var selectChange = 0;
  var selectInput = 0;
  var expectedChange = 0;
  var expectedInput = 0;

  selectSingleOneItem.focus();
  synthesizeKey("VK_DOWN", {});
  is(selectInput, expectedInput, "Down key should not fire input event when reaching end of the list.");
  is(selectChange, expectedChange, "Down key should not fire change event when reaching end of the list.");

  synthesizeKey("VK_UP", {});
  is(selectInput, expectedInput, "Up key should not fire input event when reaching top of the list.");
  is(selectChange, expectedChange, "Up key should not fire change event when reaching top of the list.");

  selectSingle.focus();
  for (var i = 1; i < selectSingle.length; i++) {
    synthesizeKey("VK_DOWN", {});

    is(selectSingle.options[i].selected, true, "Option should be selected");
    is(selectInput, ++expectedInput, "Down key should fire input event.");
    is(selectChange, ++expectedChange, "Down key should fire change event.");
  }

  // We are at the end of the list, going down should not fire change event.
  synthesizeKey("VK_DOWN", {});
  is(selectInput, expectedInput, "Down key should not fire input event when reaching end of the list.");
  is(selectChange, expectedChange, "Down key should not fire change event when reaching end of the list.");

  for (var i = selectSingle.length - 2; i >= 0; i--) {
    synthesizeKey("VK_UP", {});

    is(selectSingle.options[i].selected, true, "Option should be selected");
    is(selectInput, ++expectedInput, "Up key should fire input event.");
    is(selectChange, ++expectedChange, "Up key should fire change event.");
  }

  // We are at the top of the list, going up should not fire change event.
  synthesizeKey("VK_UP", {});
  is(selectInput, expectedInput, "Up key should not fire input event when reaching top of the list.");
  is(selectChange, expectedChange, "Up key should not fire change event when reaching top of the list.");

  selectMultipleOneItem.focus();
  synthesizeKey("VK_DOWN", {});
  is(selectInput, ++expectedInput, "Down key should fire input event when reaching end of the list.");
  is(selectChange, ++expectedChange, "Down key should fire change event when reaching end of the list.");

  synthesizeKey("VK_DOWN", {});
  is(selectInput, expectedInput, "Down key should not fire input event when reaching end of the list.");
  is(selectChange, expectedChange, "Down key should not fire change event when reaching end of the list.");

  synthesizeKey("VK_UP", {});
  is(selectInput, expectedInput, "Up key should not fire input event when reaching top of the list.");
  is(selectChange, expectedChange, "Up key should not fire change event when reaching top of the list.");

  selectMultiple.focus();
  for (var i = 0; i < selectMultiple.length; i++) {
    synthesizeKey("VK_DOWN", {});

    is(selectMultiple.options[i].selected, true, "Option should be selected");
    is(selectInput, ++expectedInput, "Down key should fire input event.");
    is(selectChange, ++expectedChange, "Down key should fire change event.");
  }

  // We are at the end of the list, going down should not fire change event.
  synthesizeKey("VK_DOWN", {});
  is(selectInput, expectedInput, "Down key should not fire input event when reaching end of the list.");
  is(selectChange, expectedChange, "Down key should not fire change event when reaching end of the list.");

  for (var i = selectMultiple.length - 2; i >= 0; i--) {
    synthesizeKey("VK_UP", {});

    is(selectMultiple.options[i].selected, true, "Option should be selected");
    is(selectInput, ++expectedInput, "Up key should fire input event.");
    is(selectChange, ++expectedChange, "Up key should fire change event.");
  }

  // We are at the top of the list, going up should not fire change event.
  synthesizeKey("VK_UP", {});
  is(selectInput, expectedInput, "Up key should not fire input event when reaching top of the list.");
  is(selectChange, expectedChange, "Up key should not fire change event when reaching top of the list.");

</script>
</pre>
</body>
</html>
